Refactoring ClientRegister
