+++ src/client/ClientForm.js
@@ -3,8 +3,6 @@ import AppDispatcher from '../AppDispatcher';
 import ActionType from '../actions/ActionType';
 import PropTypes from 'prop-types';
 //import routeMap from '../routes/RouteMap';
-import IconButton from 'material-ui/IconButton';
-import FileFileUpload from 'material-ui/svg-icons/file/file-upload';
 import Formsy from 'formsy-react';
 
 export default class ClientForm extends Component {
@@ -30,20 +28,6 @@ export default class ClientForm extends Component {
           ref={(ref) => {this.form = ref; }}
           onValidSubmit={this.submitForm}>
           {this.props.children}
-          <IconButton
-            name="id_document"
-            tooltip="Documento de Identificação"
-            touch={true}
-            tooltipPosition="top-left">
-            <FileFileUpload />
-          </IconButton>
-          <IconButton
-            name="proof_of_address"
-            tooltip="Comprovante de Residência"
-            touch={true}
-            tooltipPosition="top-right">
-            <FileFileUpload />
-          </IconButton>
         </Formsy.Form>
       </div>
     );
